+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to NetVips will be documented in this file. See [here](CHANG
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.3.0] - ???
+## [2.3.0] - 2023-03-24
 ### Added
 - Add IntPtr-based overload for `Image.NewFromMemory()` ([#177](https://github.com/kleisauke/net-vips/issues/177)).
 - Add `image.Invalidate()` to drop caches on an image, and any downstream images.